Add File Content Embedding for PDFs
This adds an option in the File Content to Embed the content (this choice is only available if you have checked the option on the File Content to make the download public - the rational here is that the embed will 'expose' the content which doesn't make sense if you weren't exposing a download of the content). Currently only PDFs have support.

FossilOrigin-Name: 703a4dd7fea456d0333bf13c603233dfaa3324c70bfdac13674a51d7c4f26281