Add Cloudflare Basic Auth Worker. Improve Summary in PhotoGenerator.
By looking at a few samples I was able to come up with a Basic Auth worker for Cloudflare that seems to be working.

Fixed a Photo import bug that was putting a single period in as the summary in some cases.

FossilOrigin-Name: c54b92588327d58ff1b911452e19fb240350ae3536f21e691ab445471f711080