Quick File Rename in File Content Editor
Provide a button available when illegal characters are detected in the file name to auto-correct the problem (note that the sites rules for filenames are more restrictive than the file systems rules). Recently working with files for a new site that had more naming problems than previous sets of files it became obvious that having to go into the file rename dialog was not the best and a quick fix was a desirable option.

FossilOrigin-Name: 5cb14b421a1af0ff3f21338278c2ffa94ac9b1d27ad9d9944c30e77ebe0c0485