Bug Fix for exception on Multiple EnsureCoreWebView2Async calls
In previous versions this seemed fine - in latest packages apparently a problem, put a guard and logging in place for now. Not sure in this case with this logic in a behaviour if it is worth looking for a better way or not.

FossilOrigin-Name: e72e8e4a0913428292ade151c4739ff65b53fd0648f865fd1d55140d9463df19