Data Structure Changes fro MainSiteFeedOn and IsDraft
MainSiteFeedOn is a way to set the date content appears on the front page and in RSS feeds - I don't want every update to trigger the post to reoccur at the top and want to preserve the CreatedOn date, but for some audiences and posts do want to have the option to push content back to the 'top'.

IsDraft was originally avoided because content will still be written to disk, but I think even a 'slightly weaker than expected' version of a 'draft' could be useful.

FossilOrigin-Name: ac17a8b0f8d83fba1e08ed45bff9d3e3065259e541e37d3509cea591c5f416ff