Style Change to Header for Larger Screens
Inspired a bit by the header style I have seen around the web with nav links to the right of the Title - nice because the screen real estate for the header is minimized which goes to the heart of the idea here which is a focus on content! Roughly the same presentation on other screen sizes.

FossilOrigin-Name: cb3cadbb068d64c34febf5a17b4dd8493da34964db550e916535a09fff4bdd86