Add the leaflet-locatecontrol to maps. Include Summary for Map Points.
Include resources in the project and in headers, modify the css to reflect the location of the svg files, update readme.
Point Json now includes the summary for a point and the summary is displayed in the point pop up.

FossilOrigin-Name: 0bb2f1fec5d213debd0507ead8147832dbcea23b51a58b7a949c4155bbc1a781